🤖 refactor: mypage Component 분리
